+++ code/src/sixsq/nuvla/server/resources/customer.clj
@@ -7,6 +7,7 @@ Customer mapping to external banking system."
     [sixsq.nuvla.auth.acl-resource :as acl-resource]
     [sixsq.nuvla.auth.utils :as auth]
     [sixsq.nuvla.db.filter.parser :as parser]
+    [sixsq.nuvla.db.impl :as db]
     [sixsq.nuvla.server.resources.common.crud :as crud]
     [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
     [sixsq.nuvla.server.resources.common.utils :as u]
@@ -17,8 +18,7 @@ Customer mapping to external banking system."
     [sixsq.nuvla.server.resources.spec.customer :as customer]
     [sixsq.nuvla.server.resources.spec.customer-related :as customer-related]
     [sixsq.nuvla.server.util.metadata :as gen-md]
-    [sixsq.nuvla.server.util.response :as r]
-    [sixsq.nuvla.db.impl :as db]))
+    [sixsq.nuvla.server.util.response :as r]))
 
 
 (def ^:const resource-type (u/ns->type *ns*))