+++ code/src/sixsq/nuvla/server/resources/deployment.clj
@@ -14,13 +14,13 @@ a container orchestration engine.
     [sixsq.nuvla.server.resources.common.utils :as u]
     [sixsq.nuvla.server.resources.configuration-nuvla :as config-nuvla]
     [sixsq.nuvla.server.resources.customer :as customer]
+    [sixsq.nuvla.server.resources.customer.utils :as customer-utils]
     [sixsq.nuvla.server.resources.deployment.utils :as dep-utils]
     [sixsq.nuvla.server.resources.event.utils :as event-utils]
-    [sixsq.nuvla.server.resources.customer.utils :as customer-utils]
+    [sixsq.nuvla.server.resources.pricing.stripe :as stripe]
     [sixsq.nuvla.server.resources.resource-metadata :as md]
     [sixsq.nuvla.server.resources.spec.deployment :as deployment-spec]
     [sixsq.nuvla.server.util.metadata :as gen-md]
-    [sixsq.nuvla.server.resources.pricing.stripe :as stripe]
     [sixsq.nuvla.server.util.response :as r]))
 
 
@@ -85,6 +85,13 @@ a container orchestration engine.
               {:name           "fetch-module"
                :uri            "fetch-module"
                :description    "fetch the deployment module href and merge it"
+               :method         "POST"
+               :input-message  "application/json"
+               :output-message "application/json"}
+
+              {:name           "check-dct"
+               :uri            "check-dct"
+               :description    "check if images are trusted"
                :method         "POST"
                :input-message  "application/json"
                :output-message "application/json"}])
@@ -195,6 +202,9 @@ a container orchestration engine.
                      (cond-> request
                              is-user? (update :body dissoc :owner :infrastructure-service
                                               :subscription-id)
+                             is-user? (update-in [:cimi-params :select] disj
+                                                 "owner" "infrastructure-service" "module/price"
+                                                 "module/license" "subscription-id")
                              new-acl (assoc-in [:body :acl] new-acl)
                              infra-id (assoc-in [:body :infrastructure-service] infra-id)))]
     (some-> subs-id stripe/retrieve-subscription (stripe/cancel-subscription {"invoice_now" true}))
@@ -237,6 +247,7 @@ a container orchestration engine.
         create-log-op       (u/action-map id :create-log)
         clone-op            (u/action-map id :clone)
         fetch-module-op     (u/action-map id :fetch-module)
+        check-dct-op        (u/action-map id :check-dct)
         upcoming-invoice-op (u/action-map id :upcoming-invoice)
         can-manage?         (a/can-manage? resource request)
         can-clone?          (a/can-view-data? resource request)]
@@ -259,6 +270,8 @@ a container orchestration engine.
             (update :operations conj fetch-module-op)
 
             can-manage? (update :operations conj upcoming-invoice-op)
+
+            can-manage? (update :operations conj check-dct-op)
 
             (not (dep-utils/can-delete? resource))
             (update :operations dep-utils/remove-delete))))
@@ -295,7 +308,7 @@ a container orchestration engine.
                              :body)]
       (when stopped?
         (dep-utils/delete-child-resources "deployment-parameter" id))
-      (dep-utils/create-job new-deployment request "start"))
+      (dep-utils/create-job new-deployment request "start_deployment"))
     (catch Exception e
       (or (ex-data e) (throw e)))))
 
@@ -308,7 +321,7 @@ a container orchestration engine.
         (dep-utils/throw-can-not-do-action dep-utils/can-stop? "stop")
         (edit-deployment request #(assoc % :state "STOPPING"))
         :body
-        (dep-utils/create-job request "stop"))
+        (dep-utils/create-job request "stop_deployment"))
     (catch Exception e
       (or (ex-data e) (throw e)))))
 
@@ -338,6 +351,17 @@ a container orchestration engine.
       (or (ex-data e) (throw e)))))
 
 
+(defmethod crud/do-action [resource-type "check-dct"]
+  [{{uuid :uuid} :params :as request}]
+  (try
+    (-> (str resource-type "/" uuid)
+        (crud/retrieve-by-id-as-admin)
+        (a/throw-cannot-manage request)
+        (dep-utils/create-job request "dct_check"))
+    (catch Exception e
+      (or (ex-data e) (throw e)))))
+
+
 (defmethod crud/do-action [resource-type "clone"]
   [{{uuid :uuid} :params :as request}]
   (try
@@ -356,7 +380,7 @@ a container orchestration engine.
         (crud/retrieve-by-id-as-admin)
         (edit-deployment request #(assoc % :state "UPDATING"))
         :body
-        (dep-utils/create-job request "update"))
+        (dep-utils/create-job request "update_deployment"))
     (catch Exception e
       (or (ex-data e) (throw e)))))